extract a package private method
